[Refactor] post 지도 관련 로직 변경
- PostMapView 분리
- PostMapDetailView ignoreSafeArea 적용
